Headless: Reduce some duplication.
We can share the screenshot code for SDL.
