+++ headless/StubHost.h
@@ -64,13 +64,21 @@ public:
 			debugOutputBuffer_.clear();
 		}
 	}
-	virtual void SetComparisonScreenshot(const std::string &filename) {}
 
+	virtual void SetComparisonScreenshot(const std::string &filename) {
+		comparisonScreenshot_ = filename;
+	}
+
+	void SendDebugScreenshot(const u8 *pixbuf, u32 w, u32 h) override;
 
 	// Unique for HeadlessHost
 	virtual void SwapBuffers() {}
 
 protected:
+	void SendOrCollectDebugOutput(const std::string &output);
+
+	std::string comparisonScreenshot_;
 	std::string debugOutputBuffer_;
 	GPUCore gpuCore_;
+	GraphicsContext *gfx_ = nullptr;
 };
